restructure/enhance
- check for files before attempting to delete
- create directories before attempting to create files in them, to
  prevent failure
- move config to parameters
- support -WhatIf and -Confirm parameters
- get Intermediary certificate from interim.pem instead of text in
  the script
- revise installation/use instructions
- [untested] Install ACMESharp, if needed

+++ GetSSL.ps1
@@ -1,93 +1,180 @@
-﻿<#
-Created by: David Nahodyl, Blue Feather 10/8/2016
-Contact: contact@bluefeathergroup.com
-Last Updated: 2/12/18
-Version: 0.6
+<#
+.SYNOPSIS
+	Get an SSL certificate from Let's Encrypt and install it on FileMaker Server.
 
-Need help? We can set this up to run on your server for you! Send an email to
-contact@bluefeathergroup.com or give a call at (770) 765-6258
+.PARAMETER Domains
+	Array of domain(s) for which you would like an SSL Certificate.
+	Let's Encrypt will peform separate validation for each of the domains,
+	so be sure that your server is reachable at all of them before
+	attempting to get a certificate. 100 domains is the max.
+
+.PARAMETER Email
+	Contact email address to your real email address so that Let's Encrypt
+	can contact you if there are any problems.
+
+.PARAMETER FMSPath
+	Path to your FileMaker Server directory, ending in a backslash. Only
+	necessary if installed in a non-default location.
+
+.NOTES
+	File Name:   GetSSL.ps1
+	Author:      David Nahodyl contact@bluefeathergroup.com, modified by Daniel Smith dan@filemaker.consulting
+	Created:     2016-10-08
+	Revised:     2018-05-17
+	Version:     0.7-DS
+
+.LINK
+	http://bluefeathergroup.com/blog/how-to-use-lets-encrypt-ssl-certificates-with-filemaker-server/
+
+.EXAMPLE
+	.\GetSSL.ps1 test.com user@test.com
+
+	Simplest call with domain to sign listed first and email second.
+
+.EXAMPLE
+	.\GetSSL.ps1 test.com, sub.example.com user@test.com
+
+	Multiple domains can be listed, separated by commas.
+
+.EXAMPLE
+	.\GetSSL.ps1 -d test.com -e user@test.com
+
+	Can use short-hand parameter names.
+
+.EXAMPLE
+	.\GetSSL.ps1 -Domain test.com -Email user@test.com
+
+	Or full parameter names.
+
+.EXAMPLE
+	.\GetSSL.ps1 test.com user@test.com -FMSPath "X:\FileMaker Server\"
+
+	Use if you installed FileMaker Server in a non-default path.
+	Must end in a backslash.
+
+.EXAMPLE
+	.\GetSSL.ps1 test.com user@test.com -Confirm
+
+	Don't ask for confirmation; use the -Confirm parameter when called from a scheduled task.
+	To have this script run silently, it must also be able to perform fmsadmin.exe without asking for username and password. There are two ways to do that:
+		1. Add a group name that is allowed to access the Admin Console and run the script as a user that belongs to the group.
+		2. Hard-code the username and password into this script. (NOT RECOMMENDED)
+
+.EXAMPLE
+	.\GetSSL.ps1 test.com user@test.com -WhatIf
+
+	Display the inputs, then exit; use to verify you passed parameters in the correct format
 #>
 
-<#  Change the domain variable to the domain/subdomain for which you would like
-	an SSL Certificate#>
-$domains = @('fms.mycompany.com');
 
-<# You can also get a certificate for multiple host name. Uncomment the line below
-and enter your domains in the array matching the example format if you'd like a
-mult-domain certificate. Let's Encrypt will peform separate validation for each
-of the domains, so be sure that your server is reachable at all of them before
-attempting to get a certificate. #>
-#$domains = @('fms.mycompany.com', 'secondaddress.mycompany.com');
+[cmdletbinding(SupportsShouldProcess=$true,ConfirmImpact='High')]
 
+Param(
+	[Parameter(Mandatory=$True,Position=1)]
+	[Alias('d')]
+	[string[]] $Domains,
 
-<# 	Change the contact email address to your real email address so that Let's Encrypt
-	can contact you if there are any problems #>
-$email = 'test@mydomain.com'
+	[Parameter(Mandatory=$True,Position=2)]
+	[Alias('e')]
+	[string] $Email,
 
-<# Enter the path to your FileMaker Server directory, ending in a backslash \ #>
-$fmsPath = 'C:\Program Files\FileMaker\FileMaker Server\'
+	[Parameter(Position=3)]
+	[Alias('p')]
+	[string] $FMSPath = 'C:\Program Files\FileMaker\FileMaker Server\'
+)
 
+<# Exit immediately on error #>
+$ErrorActionPreference = "Stop"
 
+$fmsadmin = $FMSPath + 'Database Server\fmsadmin.exe'
 
-<#
-You should not need to edit anything below this point
--------------------------------#>
-
-<# Check to make sure people changed the email address and domain #>
-#if ($email -eq('test@mydomain.com')){
-#    Write-Output 'You must enter your real email address! The script will now exit.'
-#    exit
-#}
-if ($domain -eq('fms.mydomain.com')){
-    Write-Output 'You must enter your real doamin! The script will now exit.'
-    exit
-}
-
-<# Check to make sure people changed the email address and domain #>
-if ($email -eq('test@mydomain.com')){
-    Write-Output 'You must enter your own email address! The script will now exit.'
-    exit
-}
-
-<# Check if administrator #>
 
 function Test-Administrator
 {
-    $user = [Security.Principal.WindowsIdentity]::GetCurrent();
-    (New-Object Security.Principal.WindowsPrincipal $user).IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator)
+	$user = [Security.Principal.WindowsIdentity]::GetCurrent();
+	(New-Object Security.Principal.WindowsPrincipal $user).IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator)
 }
+
+
+<# Display user input #>
+Write-Output ""
+Write-Output ('  domains:   '+($Domains -join ', '))
+Write-Output "  email:     $Email"
+Write-Output "  FMSPath:   $FMSPath"
+Write-Output ""
+
+
+<# -WhatIf parameter provided, so don't do anything; just show parameters #>
+if ([bool]$WhatIfPreference.IsPresent){
+	if (-not (Test-Administrator)){
+		Write-Output "  WARNING: This script is not running as Administrator."
+		Write-Output ""
+	}
+	if (-not(Test-Path $fmsadmin)){
+		Write-Output "  WARNING: fmsadmin could not be found at: '$fmsadmin', please check the FMSPath parameter: '$FMSPath'";
+		Write-Output ""
+	}
+	exit;
+}
+
 
 <# Check to make sure we're running as admin #>
 if (-not (Test-Administrator)){
-    Write-Output 'This script must be run as Administrator'
-    exit
+	throw 'This script must be run as Administrator'
 }
 
 
-$domainAliases = @();
-
-foreach ( $domain in $domains) {
-    $domainAliases += "$domain"+[guid]::NewGuid().ToString();
+<# validate FMSPath #>
+if (-not(Test-Path $fmsadmin)){
+	throw "fmsadmin could not be found at: '$fmsadmin', please check the FMSPath parameter: '$FMSPath'";
 }
 
-<#Install ACMESharp #>
+
+if ($ConfirmPreference -eq "High"){
+	while( -not ( ($choice= (Read-Host "If you proceed, and this script is successful, FileMaker Server service will be restarted and ALL USERS DISCONNECTED. Use the -Confirm parameter to prevent this confirmation.`n`nContinue?[y/n]")) -match "[yY]|[nN]")){ "Y or N ?"}
+	if ($choice -match "[nN]"){
+		exit;
+	}
+}
+
+
+$domainAliases = @();foreach ($domain in $Domains) {
+	if ($domain -Match ",| "){
+		throw "Domain cannot contain a comma or parameter; perhaps two domains were passed as a single string? Try removing quotes from the domains.";
+	}
+	$domainAliases += "$domain"+[guid]::NewGuid().ToString();
+}
+
+
+<# Install ACMESharp #>
+if (!(Get-Module -Listavailable -Name ACMESharp)){
+	Install-Module -Name ACMESharp, ACMESharp.Providers.IIS -AllowClobber -Confirm
+	Enable-ACMEExtensionModule -ModuleName ACMESharp.Providers.IIS
+}
 Import-Module ACMESharp;
 
+
 <# Initialize the vault to either Live or Staging#>
-
-<# Live Server #>
-Initialize-ACMEVault;
-
-<# Staging Server #>
+if (!(Get-ACMEVault)) {
+	Initialize-ACMEVault;
+}
 #Initialize-ACMEVault -BaseURI https://acme-staging.api.letsencrypt.org/
 
+
 <# Regiser contact info with LE #>
-New-ACMERegistration -Contacts mailto:$email -AcceptTos;
+New-ACMERegistration -Contacts mailto:$Email -AcceptTos;
 
 <# ACMESharp keeps creating a web.config that doesn't work, so let's delete it and make our own good one #>
-$webConfigPath = $fmsPath + 'HTTPServer\conf\.well-known\acme-challenge\web.config';
+$webConfigPath = $FMSPath + 'HTTPServer\conf\.well-known\acme-challenge\web.config';
 <# Delete the bad one #>
-Remove-Item $webConfigPath;
+if (Test-Path $webConfigPath){
+	Remove-Item $webConfigPath;
+}
+
+<# Create directory the file goes in #>
+if (-not (Test-Path (Split-Path -Path $webConfigPath -Parent))){
+	New-Item -Path (Split-Path -Path $webConfigPath -Parent) -ItemType Directory
+}
 
 <# Write a new good one #>
 ' <configuration>
@@ -96,22 +183,20 @@ Remove-Item $webConfigPath;
              <mimeMap fileExtension="." mimeType="text/plain" />
          </staticContent>
      </system.webServer>
- </configuration>' | Out-File -FilePath $webConfigPath;
+ </configuration>' | Out-File -FilePath ($webConfigPath);
 
 
 <# Loop through the array of domains and validate each one with LE #>
-
-for ( $i=0; $i -lt $domains.length; $i++ ) {
+for ( $i=0; $i -lt $Domains.length; $i++ ) {
 	<# Create a UUID alias to use for our domain request #>
-    $domain = $domains[$i];
+	$domain = $Domains[$i];
 	$domainAlias = $domainAliases[$i];
-    Write-Output "Performing challenge for $domain with alias $domainAlias";
-
+	Write-Output "Performing challenge for $domain with alias $domainAlias";
 	<#Create an entry for us to use with these requests using the alias we just generated #>
-	New-ACMEIdentifier -Dns $domain -Alias $domainAlias;
+	New-ACMEIdentifier -Dns $domain -Alias $domainAlias;
 	<# Use ACMESharp to automatically create the correct files to use for validation with LE #>
-	$response = Complete-ACMEChallenge $domainAlias -ChallengeType http-01 -Handler iis -HandlerParameters @{ WebSiteRef = 'FMWebSite'; SkipLocalWebConfig = $true } -Force;
-
+	$response = Complete-ACMEChallenge $domainAlias -ChallengeType http-01 -Handler iis -HandlerParameters @{ WebSiteRef = 'FMWebSite'; SkipLocalWebConfig = $true } -Force;
+
 	<# Sample Response
 	== Manual Challenge Handler - HTTP ==
 	  * Handle Time: [1/12/2016 1:16:34 PM]
@@ -123,17 +208,15 @@ for ( $i=0; $i -lt $domains.length; $i++ ) {
 	  * File Path: [.well-known/acme-challenge/2yRd04TwqiZTh6TWLZ1azL15QIOGaiRmx8MjAoA5QH0]
 	  * File Content: [2yRd04TwqiZTh6TWLZ1azL15QIOGaiRmx8MjAoA5QH0.H3URk7qFUvhyYzqJySfc9eM25RTDN7bN4pwil37Rgms]
 	  * MIME Type: [text/plain]------------------------------------
-	#>
+	#>
 	<# Let them know it's ready #>
-	Submit-ACMEChallenge $domainAlias -ChallengeType http-01 -Force;
+	Submit-ACMEChallenge $domainAlias -ChallengeType http-01 -Force;
 	<# Pause 10 seconds to wait for LE to validate our settings #>
 	Start-Sleep -s 10
-
 	<# Check the status #>
-	(Update-ACMEIdentifier $domainAlias -ChallengeType http-01).Challenges | Where-Object {$_.Type -eq "http-01"};
-
+	(Update-ACMEIdentifier $domainAlias -ChallengeType http-01).Challenges | Where-Object {$_.Type -eq "http-01"};
+
 	<# Good Response Sample
-
 	ChallengePart          : ACMESharp.Messages.ChallengePart
 	Challenge              : ACMESharp.ACME.HttpChallenge
 	Type                   : http-01
@@ -148,7 +231,6 @@ for ( $i=0; $i -lt $domains.length; $i++ ) {
 	HandlerCleanUpDate     :
 	SubmitDate             : 11/3/2016 12:34:48 AM
 	SubmitResponse         : {StatusCode, Headers, Links, RawContent...}
-
 	#>
 }
 
@@ -157,13 +239,17 @@ for ( $i=0; $i -lt $domains.length; $i++ ) {
 $certAlias = 'cert-'+[guid]::NewGuid().ToString();
 
 <# Ready to get the certificate #>
+Write-Output "New-ACMECertificate -----------------------------------------------------------";
 New-ACMECertificate $domainAliases[0] -Generate -AlternativeIdentifierRefs $domainAliases -Alias $certAlias;
+
+Write-Output "Submit-ACMECertificate --------------------------------------------------------";
 Submit-ACMECertificate $certAlias;
 
 <# Pause 10 seconds to wait for LE to create the certificate #>
 Start-Sleep -s 10
 
 <# Check the status $certAlias #>
+Write-Output "Update-ACMECertificate --------------------------------------------------------";
 Update-ACMECertificate $certAlias;
 
 
@@ -171,62 +257,39 @@ Update-ACMECertificate $certAlias;
 
 
 <# Export the private key #>
-$keyPath = $fmsPath + 'CStore\serverKey.pem'
-Remove-Item $keyPath;
+$keyPath = $FMSPath + 'CStore\serverKey.pem'
+if (Test-Path $keyPath){
+	Remove-Item $keyPath;
+}
 Get-ACMECertificate $certAlias -ExportKeyPEM $keyPath;
 
 <# Export the certificate #>
-$certPath = $fmsPath + 'CStore\crt.pem'
-Remove-Item $certPath;
+$certPath = $FMSPath + 'CStore\crt.pem'
+if (Test-Path $certPath){
+	Remove-Item $certPath;
+}
 Get-ACMECertificate $certAlias -ExportCertificatePEM $certPath;
 
 <# Export the Intermediary #>
-$intermPath = $fmsPath + 'CStore\interm.pem'
-Remove-Item $intermPath;
+$intermPath = $FMSPath + 'CStore\interm.pem'
+if (Test-Path $intermPath){
+	Remove-Item $intermPath;
+}
 Get-ACMECertificate $certAlias -ExportIssuerPEM $intermPath;
 
-<# cd to FMS directory to run fmsadmin commands #>
-cd $fmsPath'\Database Server\';
-
 <# Install the certificate #>
-.\fmsadmin certificate import $certPath;
+& $fmsadmin certificate import $certPath -y;
 
 <# Append the intermediary certificate to support older FMS before 15 #>
-Add-Content $fmsPath'CStore\serverCustom.pem' '
------BEGIN CERTIFICATE-----
-MIIEkjCCA3qgAwIBAgIQCgFBQgAAAVOFc2oLheynCDANBgkqhkiG9w0BAQsFADA/
-MSQwIgYDVQQKExtEaWdpdGFsIFNpZ25hdHVyZSBUcnVzdCBDby4xFzAVBgNVBAMT
-DkRTVCBSb290IENBIFgzMB4XDTE2MDMxNzE2NDA0NloXDTIxMDMxNzE2NDA0Nlow
-SjELMAkGA1UEBhMCVVMxFjAUBgNVBAoTDUxldCdzIEVuY3J5cHQxIzAhBgNVBAMT
-GkxldCdzIEVuY3J5cHQgQXV0aG9yaXR5IFgzMIIBIjANBgkqhkiG9w0BAQEFAAOC
-AQ8AMIIBCgKCAQEAnNMM8FrlLke3cl03g7NoYzDq1zUmGSXhvb418XCSL7e4S0EF
-q6meNQhY7LEqxGiHC6PjdeTm86dicbp5gWAf15Gan/PQeGdxyGkOlZHP/uaZ6WA8
-SMx+yk13EiSdRxta67nsHjcAHJyse6cF6s5K671B5TaYucv9bTyWaN8jKkKQDIZ0
-Z8h/pZq4UmEUEz9l6YKHy9v6Dlb2honzhT+Xhq+w3Brvaw2VFn3EK6BlspkENnWA
-a6xK8xuQSXgvopZPKiAlKQTGdMDQMc2PMTiVFrqoM7hD8bEfwzB/onkxEz0tNvjj
-/PIzark5McWvxI0NHWQWM6r6hCm21AvA2H3DkwIDAQABo4IBfTCCAXkwEgYDVR0T
-AQH/BAgwBgEB/wIBADAOBgNVHQ8BAf8EBAMCAYYwfwYIKwYBBQUHAQEEczBxMDIG
-CCsGAQUFBzABhiZodHRwOi8vaXNyZy50cnVzdGlkLm9jc3AuaWRlbnRydXN0LmNv
-bTA7BggrBgEFBQcwAoYvaHR0cDovL2FwcHMuaWRlbnRydXN0LmNvbS9yb290cy9k
-c3Ryb290Y2F4My5wN2MwHwYDVR0jBBgwFoAUxKexpHsscfrb4UuQdf/EFWCFiRAw
-VAYDVR0gBE0wSzAIBgZngQwBAgEwPwYLKwYBBAGC3xMBAQEwMDAuBggrBgEFBQcC
-ARYiaHR0cDovL2Nwcy5yb290LXgxLmxldHNlbmNyeXB0Lm9yZzA8BgNVHR8ENTAz
-MDGgL6AthitodHRwOi8vY3JsLmlkZW50cnVzdC5jb20vRFNUUk9PVENBWDNDUkwu
-Y3JsMB0GA1UdDgQWBBSoSmpjBH3duubRObemRWXv86jsoTANBgkqhkiG9w0BAQsF
-AAOCAQEA3TPXEfNjWDjdGBX7CVW+dla5cEilaUcne8IkCJLxWh9KEik3JHRRHGJo
-uM2VcGfl96S8TihRzZvoroed6ti6WqEBmtzw3Wodatg+VyOeph4EYpr/1wXKtx8/
-wApIvJSwtmVi4MFU5aMqrSDE6ea73Mj2tcMyo5jMd6jmeWUHK8so/joWUoHOUgwu
-X4Po1QYz+3dszkDqMp4fklxBwXRsW10KXzPMTZ+sOPAveyxindmjkW8lGy+QsRlG
-PfZ+G6Z6h7mjem0Y+iWlkYcV4PIWL1iwBi8saCbGS5jN2p8M+X+Q7UNKEkROb3N6
-KOqkqm57TH2H3eDJAkSnh6/DNFu0Qg==
------END CERTIFICATE-----'
+Add-Content $FMSPath'CStore\serverCustom.pem' (Get-Content $intermPath)
 
 <# Restart the FMS service #>
-Write-Output 'Automatically Stopping FileMaker Server'
 net stop 'FileMaker Server';
-Write-Output 'Automatically Starting FileMaker Server'
 net start 'FileMaker Server';
 
+<# Just in case server isn't configured to start automatically
+	(should add other services here, if necessary, like WPE) #>
+& $fmsadmin start server
 
 <# All done! Exit. #>
 exit;
